requirements.txt
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.2
bcrypt==3.2.0
blessed==1.18.1
botocore==1.21.36
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.8
Django==3.2.7
django-cors-headers==3.10.0
django-nose==1.4.7
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
nose==1.3.7
openapi-codec==1.3.2
packaging==21.0
paramiko==2.7.2
pathspec==0.5.9
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
simplejson==3.17.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
uritemplate==4.0.0
urllib3==1.26.6
wcwidth==0.1.9
websocket-client==0.59.0
